update Home/Grille Cipher Attack
This version passes all tests.

It tries to remove possibilities of indexes in each iteration from previous one, meaning possible indexes after a rotation should be in a set of indexes that are from the previous iteration after a rotation.

This can move the number of candidates significantly in each iteration and leaves a small number of it on the fourth round.